+++ app/src/main/java/com/techcourse/ManualHandlerMapping.java
@@ -1,7 +1,5 @@
 package com.techcourse;
 
-import com.techcourse.controller.manual.LoginController;
-import com.techcourse.controller.manual.LoginViewController;
 import com.techcourse.controller.manual.LogoutController;
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.HashMap;
@@ -21,8 +19,6 @@ public class ManualHandlerMapping implements HandlerMapping<Controller> {
     @Override
     public void initialize() {
         controllers.put("/", new ForwardController("/index.jsp"));
-        controllers.put("/login", new LoginController());
-        controllers.put("/login/view", new LoginViewController());
         controllers.put("/logout", new LogoutController());
 
         log.info("Initialized Handler Mapping!");
