--- app/src/main/java/com/techcourse/controller/manual/LoginController.java
@@ -1,38 +1,0 @@
-package com.techcourse.controller.manual;
-
-import com.techcourse.controller.UserSession;
-import com.techcourse.domain.User;
-import com.techcourse.repository.InMemoryUserRepository;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
-
-public class LoginController implements Controller {
-
-    private static final Logger log = LoggerFactory.getLogger(LoginController.class);
-
-    @Override
-    public String execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
-        if (UserSession.isLoggedIn(req.getSession())) {
-            return "redirect:/index.jsp";
-        }
-
-        return InMemoryUserRepository.findByAccount(req.getParameter("account"))
-                .map(user -> {
-                    log.info("User : {}", user);
-                    return login(req, user);
-                })
-                .orElse("redirect:/401.jsp");
-    }
-
-    private String login(final HttpServletRequest request, final User user) {
-        if (user.checkPassword(request.getParameter("password"))) {
-            final var session = request.getSession();
-            session.setAttribute(UserSession.SESSION_KEY, user);
-            return "redirect:/index.jsp";
-        }
-        return "redirect:/401.jsp";
-    }
-}
